+++ NINA.Photon.Plugin.ASA/Trigger/MLPTafterFlip.cs
@@ -241,6 +241,8 @@ namespace NINA.Photon.Plugin.ASA.MLTP
             initialized = false;
             initialTime = DateTime.MinValue;
             options.LastMLPT = DateTime.MinValue;
+            options.ActiveMLPTDurationSeconds = 0.0d;
+            options.ActiveMLPTPointCount = 0;
 
             try
             {
@@ -297,7 +299,6 @@ namespace NINA.Photon.Plugin.ASA.MLTP
                 throw new Exception("ASA MLPT model build failed");
             }
             initialTime = DateTime.Now;
-            options.LastMLPT = DateTime.Now;
 
             if (Recenter)
             {